feat: N:1 db 사용법 및 views 데코레이터 학습

+++ Web/Backend/Django/articles/views.py
@@ -1,8 +1,11 @@
 from django.shortcuts import render, redirect
-from .models import Article
-from .forms import ArticleForm
+from .models import Article, Comment
+from .forms import ArticleForm, CommentForm
 from django.contrib.auth.decorators import login_required
+from django.http import HttpRequest
+from django.views.decorators.http import require_http_methods, require_GET, require_POST, require_safe
 
+@require_safe
 def index(request):
     articles = Article.objects.all()
     context = {
@@ -10,20 +13,27 @@ def index(request):
     }
     return render(request, 'articles/index.html', context)
 
-
+@require_safe
 def detail(request, pk):
     article = Article.objects.get(pk=pk)
+    comments = article.comment_set.all()
+    comment_form = CommentForm()
     context = {
         'article': article,
+        'comments': comments,
+        'comment_form': comment_form,
     }
     return render(request, 'articles/detail.html', context)
 
 @login_required
-def create(request):
+@require_http_methods(['GET, POST'])
+def create(request: HttpRequest):
     if request.method == 'POST':
         form = ArticleForm(request.POST, request.FILES)
         if form.is_valid():
-            article = form.save()
+            article:Article = form.save(commit=False)
+            article.user = request.user
+            article.save()
             return redirect('articles:detail', article.pk)
     else:
         form = ArticleForm()
@@ -33,23 +43,52 @@ def create(request):
     return render(request, 'articles/create.html', context)
 
 @login_required
-def delete(request, pk):
+@require_POST
+def delete(request: HttpRequest, pk):
     article = Article.objects.get(pk=pk)
-    article.delete()
+    if request.user == article.user:
+        article.delete()
     return redirect('articles:index')
 
 @login_required
-def update(request, pk):
+@require_http_methods(['GET, POST'])
+def update(request: HttpRequest, pk):
     article = Article.objects.get(pk=pk)
-    if request.method == 'POST':
-        form = ArticleForm(request.POST, request.FILES, instance=article)
-        if form.is_valid():
-            form.save()
-            return redirect('articles:detail', article.pk)
+    if request.user == article.user:
+        if request.method == 'POST':
+            form = ArticleForm(request.POST, request.FILES, instance=article)
+            if form.is_valid():
+                form.save()
+                return redirect('articles:detail', article.pk)
+        else:
+            form = ArticleForm(instance=article)
     else:
-        form = ArticleForm(instance=article)
+        return redirect('articles:index')
     context = {
         'article': article,
         'form': form,
     }
     return render(request, 'articles/update.html', context)
+
+@login_required
+@require_POST
+def comment_create(request: HttpRequest, pk):
+    article = Article.objects.get(pk=pk)
+    if request.method == 'POST':
+        form = CommentForm(request.POST)
+        if form.is_valid:
+            comment: Comment = form.save(commit=False)
+            comment.article = article
+            comment.user = request.user
+            comment.save()
+    return redirect('articles:detail', pk)
+
+@login_required
+@require_POST
+def comment_delete(request: HttpRequest, article_pk, comment_pk):
+    comment = Comment.objects.get(pk=comment_pk)
+    if comment.user == request.user:
+        comment.delete()
+    return redirect('articles:detail', article_pk)
+
+
